Reextracted steps to separate files
