+++ app/src/main/java/app/gamenative/utils/preInstallSteps/PreInstallStep.kt
@@ -1,0 +1,24 @@
+package app.gamenative.utils
+
+import app.gamenative.data.GameSource
+import app.gamenative.enums.Marker
+import com.winlator.container.Container
+import java.io.File
+
+interface PreInstallStep {
+    val marker: Marker
+
+    fun appliesTo(
+        container: Container,
+        gameSource: GameSource,
+        gameDirPath: String,
+    ): Boolean
+
+    fun buildCommand(
+        container: Container,
+        appId: String,
+        gameSource: GameSource,
+        gameDir: File,
+        gameDirPath: String,
+    ): String?
+}
